pyproject.toml
[project]
name = "razorpay_integration"
version = "0.0.1"
description = "Razorpay Integration for ERPNext"
authors = [
    { name = "Rejith R", email = "rejithr1995@gmail.com" }
]
license = { text = "MIT" }
dependencies = [
    "frappe",
    "razorpay==1.3.0",
    "requests==2.28.1",
    "qrcode==7.3.1",
    "Pillow==9.0.0"
]
requires-python = ">=3.10"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/yourusername/razorpay_integration"
Repository = "https://github.com/yourusername/razorpay_integration.git"